+++ view.php
@@ -44,47 +44,106 @@ $stepid = optional_param('stepid', null, PARAM_INT);
 $triggerid = optional_param('triggerid', null, PARAM_INT);
 $courseid = optional_param('courseid', null, PARAM_INT);
 
-$PAGE->set_title(get_string('viewheading', 'tool_lifecycle'));
-$PAGE->set_heading(get_string('viewheading', 'tool_lifecycle'));
+// Bulk edit params.
+$bulkedit = optional_param('bulkedit', 0, PARAM_INT);
+$bulkactions = optional_param_array('bulkactions', [], PARAM_TEXT);
+$reportparam = optional_param_array('report', [], PARAM_RAW);
 
 $controller = new view_controller();
 
-if ($action !== null && $processid !== null && $stepid !== null) {
+$report = [];
+if ($bulkactions) {
     require_sesskey();
-    $controller->handle_interaction($action, $processid, $stepid);
-    exit;
-} else if ($triggerid !== null && $courseid !== null) {
-    require_sesskey();
-    $controller->handle_trigger($triggerid, $courseid);
-    exit;
+    foreach ($bulkactions as $action) {
+        $item = explode('_', $action);
+        if (count($item) == 2) {
+            $triggerid = $item[0];
+            $courseid = $item[1];
+            $course = get_course($courseid);
+            $coursename = get_course_display_name_for_list($course);
+            if ($rc = $controller->handle_trigger($triggerid, $courseid)){
+                $report[] = \html_writer::div($coursename.": ".$rc, 'alert alert-danger');
+            } else {
+                $successmsg = get_string('manual_trigger_success', 'tool_lifecycle');
+                $report[] = \html_writer::div($coursename.": ".$successmsg, 'alert alert-success');
+            }
+        } else if (count($item) == 3) {
+            $action = $item[0];
+            $processid = $item[1];
+            $courseid = $DB->get_field_sql(
+                "SELECT courseid FROM {tool_lifecycle_process} WHERE id = :pid", ['pid' => $processid]);
+            $course = get_course($courseid);
+            $coursename = get_course_display_name_for_list($course);
+            $stepid = $item[2];
+            if ($controller->handle_interaction($action, $processid, $stepid)) {
+                $successmsg = get_string('interaction_success', 'tool_lifecycle');
+                $report[] = \html_writer::div($coursename.": ".$successmsg, 'alert alert-success');
+            }
+        }
+    }
+} else {
+    if ($action !== null && $processid !== null && $stepid !== null) {
+        require_sesskey();
+        $courseid = $DB->get_field_sql(
+            "SELECT courseid FROM {tool_lifecycle_process} WHERE id = :pid", ['pid' => $processid]);
+        $course = get_course($courseid);
+        $coursename = get_course_display_name_for_list($course);
+        if ($controller->handle_interaction($action, $processid, $stepid)) {
+            $successmsg = get_string('interaction_success', 'tool_lifecycle');
+            $report[] = \html_writer::div($coursename.": ".$successmsg, 'alert alert-success');
+        }
+    } else if ($triggerid !== null && $courseid !== null) {
+        require_sesskey();
+        $course = get_course($courseid);
+        $coursename = get_course_display_name_for_list($course);
+        if ($rc = $controller->handle_trigger($triggerid, $courseid)){
+            $report[] = \html_writer::div($coursename.": ".$rc, 'alert alert-danger');
+        } else {
+            $successmsg = get_string('manual_trigger_success', 'tool_lifecycle');
+            $report[] = \html_writer::div($coursename.": ".$successmsg, 'alert alert-success');
+        }
+    }
 }
+if ($report) {
+    $redirecturl = new moodle_url($PAGE->url, ['report' => $report, 'bulkedit' => $bulkedit]);
+    redirect($redirecturl);
+}
+
+$PAGE->set_title(get_string('viewheading', 'tool_lifecycle'));
+$PAGE->set_heading(get_string('viewheading', 'tool_lifecycle'));
 
 $renderer = $PAGE->get_renderer('tool_lifecycle');
 
 echo $renderer->header();
 
-$mform = new form_courses_filter();
+$filterform = new form_courses_filter();
 
 // Cache handling.
 $cache = cache::make('tool_lifecycle', 'mformdata');
-if ($mform->is_cancelled()) {
+if ($filterform->is_cancelled()) {
     $cache->delete('viewcourses_filter');
     redirect($PAGE->url);
-} else if ($data = $mform->get_data()) {
+} else if ($data = $filterform->get_data()) {
     $cache->set('viewcourses_filter', $data);
 } else {
     $data = $cache->get('viewcourses_filter');
     if ($data) {
-        $mform->set_data($data);
+        $filterform->set_data($data);
+    }
+}
+
+if ($reportparam) {
+    foreach ($reportparam as $message) {
+        echo $message;
     }
 }
 
 echo '<br>';
 
-$mform->display();
+$filterform->display();
 
 echo '<br>';
 
-$controller->handle_view($renderer, $data);
+$controller->handle_view($renderer, $data, $bulkedit);
 
 echo $renderer->footer();